Prepare for fix to https://github.com/flutter/flutter/issues/109339. (#1121)
Currently, in some circumstances where a subclasses of
`PlatformInterface` erroneously uses `implements` rather than
`extends`, a `NoSuchMethodError` will be thrown (in spite of the
documentation at
https://pub.dev/documentation/plugin_platform_interface/latest/plugin_platform_interface/PlatformInterface/verify.html
claiming that `AssertionError` will be thrown).

After https://github.com/flutter/flutter/issues/109339 is fixed, the
correct type of exception will be thrown.  To avoid a test breakage in
`geolocator_platform_interface_test.dart` when the fix happens, we
need to modify the test so that it doesn't care what kind of exception
is thrown.