platform: Correctly handle integer-like numbers from JSON (#1531)
In JSON, an allowed representation for doubles without a fractional
part like `1.0` is actually `1`. This is, for example, the behavior of
[JSONEncoder][1].

When using `jsonDecode`, a `0` or `1` is deserialized to an `int`, which
would then cause `Position.fromMap` to crash. Instead, use `toDouble()` to
ensure that the value is converted to a `double`.

[1]: https://developer.apple.com/documentation/foundation/jsonencoder