MRS-3080 Add public method to force stop the location service
There are open issues in the geolocator package (https://github.com/Baseflow/flutter-geolocator/issues/1395 and https://github.com/Baseflow/flutter-geolocator/issues/1682) which prevents the location service from stopping even if the app is killed or all subscriptions are cancelled on the client side.

To tackle this we need to introduce this public method to force stop the location service when needed.